+++ tests/test_crossover.py
@@ -10,6 +10,9 @@ import leap_ec.ops as ops
 import leap_ec.statistical_helpers as stat
 
 
+##############################
+# Tests for uniform_crossover()
+##############################
 def test_uniform_crossover():
     pop = [Individual([0, 0]),
            Individual([1, 1])]
@@ -28,6 +31,71 @@ def test_uniform_crossover():
     # Note because we didn't clone the selected individuals, *the original population was changed*.
     assert pop[0].genome == [1,1]
     assert pop[1].genome == [0,0]
+
+
+@pytest.mark.stochastic
+def test_uniform_crossover_probability1():
+    """If we perform uniform rossover with a probabilty of 0.0, then the individuals will always be unmodified."""
+    N = 20
+    unmodified_count = 0
+
+    for i in range(N):
+
+        pop = [Individual([0, 0]),
+               Individual([1, 1])]
+        i = ops.naive_cyclic_selection(pop)
+        new_pop = list(itertools.islice(ops.uniform_crossover(i, p_xover=0.0), 2))
+
+        if new_pop[0].genome == [0, 0] and new_pop[1].genome == [1, 1]:
+            unmodified_count += 1
+
+    assert(unmodified_count == N)
+
+
+@pytest.mark.stochastic
+def test_n_ary_crossover_probability2():
+    """If we perform uniform crossover with a probabilty of 1.0, then we should see genes swapped
+    by default with probability 0.2."""
+    N = 1000
+    observed_dist = {'Unmodified': 0, 'Only left swapped': 0, 'Only right swapped': 0, 'Both swapped': 0 }
+
+    # Run crossover N times on a fixed pair of two-gene individuals
+    for i in range(N):
+
+        pop = [Individual([0, 0]),
+               Individual([1, 1])]
+        i = ops.naive_cyclic_selection(pop)
+        new_pop = list(itertools.islice(ops.uniform_crossover(i, p_xover=1.0), 2))
+
+        # There are four possible outcomes, which we will count the occurence of
+        if new_pop[0].genome == [0, 0] and new_pop[1].genome == [1, 1]:
+            observed_dist['Unmodified'] += 1
+        elif new_pop[0].genome == [1, 0] and new_pop[1].genome == [0, 1]:
+            observed_dist['Only left swapped'] += 1
+        elif new_pop[0].genome == [0, 1] and new_pop[1].genome == [1, 0]:
+            observed_dist['Only right swapped'] += 1
+        elif new_pop[0].genome == [1, 1] and new_pop[1].genome == [0, 0]:
+            observed_dist['Both swapped'] += 1
+        else:
+            assert(False)
+
+    assert(N == sum(observed_dist.values()))
+
+    p = 0.01
+    p_swap = 0.2
+    # This is the count we expect to see of each combination
+    # Each locus swaps with p_swap.
+    expected_dist = {
+        'Unmodified': int((1-p_swap)*(1-p_swap)*N),
+        'Only left swapped': int(p_swap*(1-p_swap)*N),
+        'Only right swapped': int((1-p_swap)*p_swap*N),
+        'Both swapped': int(p_swap**2 * N)
+    }
+
+    # Use a χ-squared test to see if our experiment matches what we expect
+    assert(stat.stochastic_equals(expected_dist, observed_dist, p=p))
+
+
 
 # These tests are now moot given the advent of variable length segments.
 
@@ -57,6 +125,9 @@ def test_uniform_crossover():
 #         new_pop = list(itertools.islice(ops.n_ary_crossover(i), 2))
 
 
+##############################
+# Tests for n_ary_crossover()
+##############################
 def test_n_ary_crossover_bad_crossover_points():
     """ Test assertions for having more crossover points than genome length """
     pop = [Individual([0, 0]),
@@ -76,7 +147,7 @@ def test_n_ary_crossover():
 
     i = ops.naive_cyclic_selection(pop)
 
-    new_pop = list(itertools.islice(ops.n_ary_crossover(i), 2))
+    new_pop = list(itertools.islice(ops.n_ary_crossover(i, num_points=1), 2))
 
     # Given that there are only two genes, one [0,0] and the other [1,1] and a single crossover point, and that the
     # only two valid crossover points are 0 or 1, then there are two possible valid states for offspring with single
@@ -96,7 +167,7 @@ def test_n_ary_crossover_probability():
         pop = [Individual([0, 0]),
                Individual([1, 1])]
         i = ops.naive_cyclic_selection(pop)
-        new_pop = list(itertools.islice(ops.n_ary_crossover(i, p=0.5), 2))
+        new_pop = list(itertools.islice(ops.n_ary_crossover(i, num_points=1, p=0.5), 2))
 
         if new_pop[0].genome == [0, 0] and new_pop[1].genome == [1, 1]:
             unmodified_count += 1